+++ controllers/adminControllers/portfolioController.js
@@ -15,28 +15,33 @@ async function indexDELETE(req, res){
   let AID = cookie.get('AID', {signed:true});
 
 
-  if(!AID)
-  {
-      res.render('adminViews/login')
-  }
-  else{
-      user.findById(`${AID}`)
-      .then(result=>{
-        if(result != null){
-          portfolioModel.findByIdAndDelete(`${req.params.id}`)
-          .then(result=>{
-            fs.unlinkSync(path.normalize(__dirname + `/../../public/img/portfolio/${req.params.id}.jpg`));
-          });
-          res.redirect('/admin/portfolio');
+  try{
+      if(!AID)
+    {
+        res.render('adminViews/login')
+    }
+    else{
+        user.findById(`${AID}`)
+        .then(result=>{
+          if(result != null){
+            portfolioModel.findByIdAndDelete(`${req.params.id}`)
+            .then(result=>{
+              if(fs.existsSync(__dirname + `/../../public/img/portfolio/${req.params.id}.jpg`)) fs.unlinkSync(path.normalize(__dirname + `/../../public/img/portfolio/${req.params.id}.jpg`));
+            });
+            res.redirect('/admin/portfolio');
 
-              }
-              else{
-                res.redirect('/');
-              }
-      })
-      .catch(err =>{
-          console.log(err);
-      })
+                }
+                else{
+                  res.redirect('/');
+                }
+        })
+        .catch(err =>{
+            console.log(err);
+        })
+    }
+  }catch(e){
+    console.log(e);
+    res.redirect('/admin');
   }
 
 }
